Extract handler to auth
