+++ Cargo.toml
@@ -63,6 +63,7 @@ chrono = { version = ">=0.4.26, <0.5", features = [
 data-encoding = ">=2.0, <3"
 ed25519-dalek = { version = ">=2.1.1, <=2.2.0" }
 futures = ">=0.3.31, <0.4"
+http = ">=1.4.0, <2.0"
 js-sys = { version = ">=0.3.72, <0.4" }
 mockall = { version = ">=0.13.1, <0.15" }
 proc-macro2 = ">=1.0.89, <2"
