+++ crates/bitwarden-auth/Cargo.toml
@@ -26,6 +26,7 @@ uniffi = ["bitwarden-core/uniffi", "dep:uniffi"] # Uniffi bindings
 
 # Note: dependencies must be alphabetized to pass the cargo sort check in the CI pipeline.
 [dependencies]
+async-trait = { workspace = true }
 bitwarden-api-api = { workspace = true }
 bitwarden-api-key-connector = { workspace = true }
 bitwarden-core = { workspace = true, features = ["internal"] }
@@ -33,6 +34,7 @@ bitwarden-crypto = { workspace = true }
 bitwarden-encoding = { workspace = true }
 bitwarden-error = { workspace = true }
 chrono = { workspace = true }
+http = { workspace = true }
 reqwest = { workspace = true }
 reqwest-middleware = { workspace = true }
 serde = { workspace = true }
