+++ crates/bitwarden-auth/src/renew/mod.rs
@@ -1,0 +1,5 @@
+//! Token renewal module.
+
+mod token_handler;
+
+pub use token_handler::AuthTokenHandler;
