+++ crates/bitwarden-core/Cargo.toml
@@ -56,7 +56,7 @@ bitwarden-uuid = { workspace = true }
 chrono = { workspace = true, features = ["std"] }
 # We don't use this directly (it's used by rand), but we need it here to enable WASM support
 getrandom = { version = ">=0.2.9, <0.3", features = ["js"] }
-http = ">=1.4.0, <2.0"
+http = { workspace = true }
 rand = ">=0.8.5, <0.9"
 reqwest = { workspace = true }
 reqwest-middleware = { workspace = true }
